+++ changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Version 1.1.0 (Tuesday 17 Jan 2023)
+
+- Add video in page only if video exists
+- Check if video exists to download
+- Cleanup the page generation HTML
+
+---
+
 ## Version 1.0.0 (Tuesday 17 Jan 2023)
 
 - Added page generation